Add skeleton for config options
