+++ FlySight/active_control.c
@@ -24,7 +24,6 @@
 #include <stdbool.h>
 
 #include "main.h"
-#include "activelook.h"
 #include "app_common.h"
 #include "audio_control.h"
 #include "baro.h"
@@ -227,9 +226,6 @@ void FS_ActiveControl_DataReady_Callback(void)
 
 	// Update BLE characteristic
 	Custom_GNSS_Update(data);
-
-	// Update ActiveLook glasses
-	FS_ActiveLook_GNSS_Update(data);
 
 	hasFix = (data->gpsFix == 3);
 }
