+++ FlySight/active_mode.c
@@ -48,9 +48,6 @@ void FS_ActiveMode_Init(void)
 	uint8_t enable_flags;
 	bool isSystemHealthy = true;
 
-	/* Initialize ActiveLook interface */
-	FS_ActiveLook_Init();
-
 	/* Initialize FatFS */
 	FS_ResourceManager_RequestResource(FS_RESOURCE_FATFS);
 
@@ -71,6 +68,12 @@ void FS_ActiveMode_Init(void)
 	if (f_chdir("/config") == FR_OK)
 	{
 		FS_Config_Read(FS_State_Get()->config_filename);
+	}
+
+	if (FS_Config_Get()->al_mode != 0)
+	{
+		/* Initialize ActiveLook interface */
+		FS_ActiveLook_Init();
 	}
 
 	if (FS_Config_Get()->enable_logging)
@@ -208,8 +211,11 @@ void FS_ActiveMode_Init(void)
 
 void FS_ActiveMode_DeInit(void)
 {
-	/* De-initialize ActiveLook interface */
-	FS_ActiveLook_DeInit();
+	if (FS_Config_Get()->al_mode != 0)
+	{
+		/* De-initialize ActiveLook interface */
+		FS_ActiveLook_DeInit();
+	}
 
 	/* Disable controller */
 	FS_ActiveControl_DeInit();
